pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "consumer-product-recalls"
version = "0.1.0"
description = "ETL pipeline for US federal consumer product recall data"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.13,<2",
    "httpx>=0.27,<1",
    "tenacity>=8.3,<9",
    "structlog>=24.2,<25",
    "typer>=0.12,<1",
    "dbt-core>=1.8,<2",
    "dbt-postgres>=1.8,<2",
    "boto3>=1.34,<2",
]

[project.scripts]
recalls = "src.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.2,<9",
    "pytest-vcr>=1.0,<2",
    "pytest-cov>=5.0,<6",
    "respx>=0.21,<1",
    "vcrpy>=6.0,<7",
    "pyyaml>=6.0,<7",
    "ruff>=0.8,<1",
    "pyright>=1.1.389,<2",
    "pre-commit>=3.7,<4",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=85"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.flake8-type-checking]
# Pydantic and pydantic-settings resolve annotations at runtime; their imports
# must not be moved to TYPE_CHECKING blocks.
runtime-evaluated-base-classes = ["pydantic.BaseModel", "pydantic_settings.BaseSettings"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
